# FP.dev/contracts/openapi.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DoSelect.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5126/"
    }
  ],
  "paths": {
...
    "/api/v1/admin/returns": {
      "get": {
        "tags": [
          "AdminReturns"
        ],
        "parameters": [
          {
            "name": "Statuses",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ReturnRequestStatus"
              }
            }
          },
          {
            "name": "ReasonCodes",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "To",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfAdminReturnSummaryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfAdminReturnSummaryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfAdminReturnSummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/returns/{id}": {
      "get": {
        "tags": [
          "AdminReturns"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReturnDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReturnDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReturnDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/returns/{id}/actions/review": {
      "post": {
        "tags": [
          "AdminReturns"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveReturnRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveReturnRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveReturnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnRequestDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnRequestDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnRequestDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/returns/{id}/actions/receive": {
      "post": {
        "tags": [
          "AdminReturns"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiveReturnRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiveReturnRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiveReturnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnRequestDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnRequestDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnRequestDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/returns/{id}/actions/inspect": {
      "post": {
        "tags": [
          "AdminReturns"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectReturnRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectReturnRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InspectReturnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnRequestDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnRequestDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnRequestDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/returns/{id}/actions/extend-shipment-deadline": {
      "post": {
        "tags": [
          "AdminReturns"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendShipmentDeadlineRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendShipmentDeadlineRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendShipmentDeadlineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnRequestDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnRequestDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnRequestDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/returns/{id}/shipment": {
      "get": {
        "tags": [
          "AdminReturns"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnShipmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnShipmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnShipmentDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AdminReturns"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReturnShipmentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReturnShipmentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReturnShipmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnShipmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnShipmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnShipmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/returns/{id}/shipment/events": {
      "post": {
        "tags": [
          "AdminReturns"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendReturnShipmentEventRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendReturnShipmentEventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AppendReturnShipmentEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnShipmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnShipmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnShipmentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/returns/{id}": {
      "get": {
        "tags": [
          "Returns"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnRequestDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnRequestDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnRequestDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminReturnDetailDto": {
        "required": [
          "return",
          "inspections",
          "refundableItemsPreview",
          "history",
          "availableActions"
        ],
        "type": "object",
        "properties": {
          "return": {
            "$ref": "#/components/schemas/ReturnRequestDto"
          },
          "inspections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReturnInspectionDto"
            }
          },
          "refundableItemsPreview": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RefundableItemPreviewDto"
            }
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReturnHistoryEntryDto"
            }
          },
          "availableActions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AdminReturnSummaryDto": {
        "required": [
          "publicId",
          "returnNumber",
          "orderPublicId",
          "orderNumber",
          "status",
          "priority",
          "itemCount",
          "requestedAtUtc",
          "returnShipmentDueAtUtc",
          "needsAttention",
          "rowVersion"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string",
            "format": "uuid"
          },
          "returnNumber": {
            "type": "string"
          },
          "orderPublicId": {
            "type": "string",
            "format": "uuid"
          },
          "orderNumber": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ReturnRequestStatus"
          },
          "priority": {
            "$ref": "#/components/schemas/CasePriority"
          },
          "itemCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "requestedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "returnShipmentDueAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "needsAttention": {
            "type": "boolean"
          },
          "rowVersion": {
            "type": "string",
            "format": "byte"
          }
        }
      },
...
      "AppendReturnShipmentEventRequest": {
        "required": [
          "source",
          "externalEventId",
          "eventType",
          "occurredAtUtc"
        ],
        "type": "object",
        "properties": {
          "source": {
            "maxLength": 32,
            "minLength": 1,
            "type": "string"
          },
          "externalEventId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "eventType": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "occurredAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApproveReturnRequest": {
        "required": [
          "approved",
          "items",
          "reasonCode",
          "returnRowVersion"
        ],
        "type": "object",
        "properties": {
          "approved": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApproveReturnItemLine"
            }
          },
          "reasonCode": {
            "maxLength": 64,
            "minLength": 1,
            "type": "string"
          },
          "note": {
            "maxLength": 500,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "assemblyFeeDisposition": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AssemblyFeeDisposition"
              }
            ]
          },
          "returnShippingCost": {
            "maximum": 79228162514264337593543950335,
            "minimum": 0,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "returnRowVersion": {
            "type": "string",
            "format": "byte"
          }
        }
      },
      "AssemblyFeeDisposition": {
        "enum": [
          "notApplicable",
          "notStarted",
          "merchantCancelled",
          "assemblyFault",
          "merchantFaultWholeUnit",
          "completedPartialReturn",
          null
        ]
      },
...
      "CasePriority": {
        "enum": [
          "low",
          "normal",
          "high",
          "urgent"
        ]
      },
...
      "CreateReturnShipmentRequest": {
        "required": [
          "method",
          "returnRowVersion"
        ],
        "type": "object",
        "properties": {
          "method": {
            "$ref": "#/components/schemas/ReturnShipmentMethod"
          },
          "carrierCode": {
            "maxLength": 32,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "recipientName": {
            "maxLength": 100,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "recipientPhone": {
            "maxLength": 30,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "postalCode": {
            "maxLength": 10,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "addressLine": {
            "maxLength": 200,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "storeCode": {
            "maxLength": 50,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "storeName": {
            "maxLength": 100,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "returnRowVersion": {
            "type": "string",
            "format": "byte"
          }
        }
      },
...
      "ExtendShipmentDeadlineRequest": {
        "required": [
          "reasonCode",
          "returnRowVersion"
        ],
        "type": "object",
        "properties": {
          "reasonCode": {
            "maxLength": 64,
            "minLength": 1,
            "type": "string"
          },
          "returnRowVersion": {
            "type": "string",
            "format": "byte"
          }
        }
      },
...
      "InspectReturnRequest": {
        "required": [
          "items",
          "returnRowVersion"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InspectReturnItemLine"
            }
          },
          "returnRowVersion": {
            "type": "string",
            "format": "byte"
          },
          "assemblyFeeDisposition": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AssemblyFeeDisposition"
              }
            ]
          },
          "returnShippingCost": {
            "maximum": 79228162514264337593543950335,
            "minimum": 0,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "PageResultOfAdminReturnSummaryDto": {
        "required": [
          "items",
          "pageNumber",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminReturnSummaryDto"
            }
          },
          "pageNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "ReceiveReturnRequest": {
        "required": [
          "returnRowVersion"
        ],
        "type": "object",
        "properties": {
          "note": {
            "maxLength": 500,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "returnRowVersion": {
            "type": "string",
            "format": "byte"
          }
        }
      },
...
      "ReturnRequestDto": {
        "required": [
          "publicId",
          "returnNumber",
          "orderPublicId",
          "orderNumber",
          "status",
          "priority",
          "reasonCode",
          "description",
          "items",
          "attachments",
          "requestedAtUtc",
          "approvedAtUtc",
          "receivedAtUtc",
          "closedAtUtc",
          "returnShipmentDueAtUtc",
          "shipmentDeadlineExtended",
          "shipment",
          "availableActions",
          "rowVersion"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string",
            "format": "uuid"
          },
          "returnNumber": {
            "type": "string"
          },
          "orderPublicId": {
            "type": "string",
            "format": "uuid"
          },
          "orderNumber": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ReturnRequestStatus"
          },
          "priority": {
            "$ref": "#/components/schemas/CasePriority"
          },
          "reasonCode": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReturnItemDto"
            }
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReturnAttachmentDto"
            }
          },
          "requestedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "approvedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "receivedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "closedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "returnShipmentDueAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "shipmentDeadlineExtended": {
            "type": "boolean"
          },
          "shipment": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReturnShipmentDto"
              }
            ]
          },
          "availableActions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rowVersion": {
            "type": "string",
            "format": "byte"
          }
        }
      },
      "ReturnRequestStatus": {
        "enum": [
          "requested",
          "underReview",
          "approved",
          "awaitingShipment",
          "inTransit",
          "received",
          "inspecting",
          "awaitingRefund",
          "completed",
          "rejected",
          "cancelled"
        ]
      },
      "ReturnShipmentDto": {
        "required": [
          "publicId",
          "shipmentNumber",
          "method",
          "status",
          "carrierCode",
          "trackingNumber",
          "maskedRecipientName",
          "maskedRecipientPhone",
          "maskedAddress",
          "storeCode",
          "storeName",
          "shippedAtUtc",
          "receivedAtUtc",
          "events",
          "rowVersion"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string",
            "format": "uuid"
          },
          "shipmentNumber": {
            "type": "string"
          },
          "method": {
            "$ref": "#/components/schemas/ReturnShipmentMethod"
          },
          "status": {
            "$ref": "#/components/schemas/ReturnShipmentStatus"
          },
          "carrierCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "trackingNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "maskedRecipientName": {
            "type": [
              "null",
              "string"
            ]
          },
          "maskedRecipientPhone": {
            "type": [
              "null",
              "string"
            ]
          },
          "maskedAddress": {
            "type": [
              "null",
              "string"
            ]
          },
          "storeCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "storeName": {
            "type": [
              "null",
              "string"
            ]
          },
          "shippedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "receivedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReturnShipmentEventSummaryDto"
            }
          },
          "rowVersion": {
            "type": "string",
            "format": "byte"
          }
        }
      },
...
    }
  },
...
}